Added `googleAndThen` and `googleSearchAndThen` in `google.sh`, the latter of which I used to refactor `yt2mp3` and `viki`.
`googleAndThen` uses the new `--url-output` option I added to `googler`
to create a named pipe that `googler` writes the selected urls to,
read from that pipe concurrently, and run a command for each url.

`googleSearchAndThen` is mostly factored out from `viki`.
It supports `url`, `code`, and `search` subcommands.
It takes a url prefix and url handler function as arguments
and calls the url handler to process each url.
If `url` is given, then it just calls the url handler directly.
If `code` is given, it prefixes the code with the prefix and calls the url handler.
If `search` is given, it runs `googleAndThen` so that the url handler
is called for each url selected in `googler`.

I then reimplemented `viki` using `googleSearchAndThen`,
as well as `yt2mp3`, so that it now can search from the command line as well.