Improved `imports.mjs` to directly handle `genConfig`.
`imports.mjs` now implements `genConfig` itself,
which it can do faster, and is easier to check the generated file's
modification time, so it will re-generate the file if it's more than a day old.
Normal files and generated files of the same name are also now allowed,
i.e., "hello.sh" and "hello.generate.sh" are both imported by "hello".