Added `cleanMusicFileNames`, which removes redundant "audio", "lyric", or "official" in music filenames (unfinished).
It does this using `nomino` for undo-able file renaming and `cleanMusicFileNames.js` for the actual complex renaming.
JS `cleanFileNameOnce` is not quite done, however.

`cleanMusicFileNames` only generates `rename.json` in $ONE_MUSIC.
`cleanMusicFileNames run` does the actual renaming.

`cleanMusicFileNames.js` is passed the rename filename,
which it appends the new JSON to.